migrate bigquery to output state
